+++ chia/_tests/util/test_full_block_utils.py
@@ -27,14 +27,14 @@ from chia_rs.sized_bytes import bytes32
 from chia_rs.sized_ints import uint8, uint32, uint64, uint128
 
 from chia._tests.util.benchmarks import rand_bytes, rand_g1, rand_g2, rand_hash, rand_vdf, rand_vdf_proof, rewards
-from chia.types.blockchain_format.serialized_program import SerializedProgram
-from chia.types.blockchain_format.vdf import VDFInfo, VDFProof
-from chia.util.full_block_utils import (
+from chia.full_node.full_block_utils import (
     block_info_from_block,
     generator_from_block,
     get_height_and_tx_status_from_block,
     header_block_from_block,
 )
+from chia.types.blockchain_format.serialized_program import SerializedProgram
+from chia.types.blockchain_format.vdf import VDFInfo, VDFProof
 from chia.util.generator_tools import get_block_header
 
 test_g2s: list[G2Element] = [rand_g2() for _ in range(10)]
