+++ tools/analyze-chain.py
@@ -24,10 +24,10 @@ from chia_rs import (
 from chia_rs.sized_bytes import bytes32
 
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
+from chia.full_node.full_block_utils import block_info_from_block, generator_from_block
 from chia.types.block_protocol import BlockInfo
 from chia.types.blockchain_format.serialized_program import SerializedProgram
 from chia.util.condition_tools import pkm_pairs
-from chia.util.full_block_utils import block_info_from_block, generator_from_block
 
 
 # returns an optional error code and an optional SpendBundleConditions (from chia_rs)
